requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2023.5.7
charset-normalizer==2.1.1
discord.py==2.1.1
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pi==0.1.2
pymongo==4.3.3
python-dotenv==0.21.1
requests==2.31.0
supervisor==4.2.5
urllib3==1.26.6
yarl==1.8.2
setuptools==68.2.0